Add graphQL sample application
This commit adds a sample application to the repository.
This application should not be published as part of the build.

In this application, we're trying to showcase the various features and
use cases with our Spring graphQL integration. We're mixing here data
fetchers backed by datastores or remote hypermedia APIs.

Closes gh-15