Refactor input argument binding
Prior to this commit, `@Argument`-annotated controller method parameters
were bound from the environment by serializing the environment argument
map to JSON and deserialized back to the target parameter type.

This allowed to cover a large spectrum of cases but showed some
limitations: performance could be improved and custom scalars could be
overridden in the process.

This commit removes the (de)serialization process and instead:

* runs through the environment argument map to collect all properties in
  a `MutablePropertyValues` map
* instantiate the target type if necessary
* binds properties to the target type using a `DataBinder`

With this change, scalar types should not be erased and arguments now
support a wider range of Collection types.

Closes gh-122