Correct handling of `complete` WebSocket message
This `handleConnectionCompletion` method on `WebSocketInterceptor`
method handles the `complete` GraphQL over WebSocket message which is
defined as canceling an individual subscription stream. The method was
incorrectly modeled as a notification for the closing of the whole
connection.

The method has been renamed, Javadoc clarified, and the subscription
id passed in.

See gh-276