Remove built-in ThreadLocalAccessor
The ThreadLocalAccessor in 1.0 was intended as a temporary solution
that is now replaced by a similar contract in the
`io.micrometer:context-propagation` library.

Unfortunately, there is no way to adapt one to the other, but
it is trivial to implement the new contract and register it either with
`ContextRegistry#getInstance` on startup or through the
`ServiceLoader` mechanism.

See gh-459