Document GraalVM Native support
This commit documents the GraalVM Native image support in Spring for
GraphQL and lists relevant pointers to the Spring Framework and Spring
Boot documentations.

This also explains how developers are expected to provide reachability
hints for the cases where the AOT processing is not able to discover the
relevant types.

Closes gh-581