support returning dataloader from suspend function
Rudimentary implementation to support returning a `CompletableFuture` from a suspend fun.

`CoroutinesUtils.invokeSuspendingFunction` wraps the return value of the function in a `Mono` (or `Flux`). But it also does this when a `CompletableFuture` is returned, and thus results in a `Mono<CompletableFuture<?>>`, which isn't captured by graphql-java, and thus the dataloader is never dispatched.

By unwrapping the future, and _converting_ it to a mono (as opposed to wrapping), the dataloader seems to be dispatched correctly.