Improve logic to resolve pagination arguments
The presence of either "after" or "first" leads to forward pagination.
Else if either "before" or "last" leads to backward pagination.
Fall back on forward if none are provided at all.

In addition, a small adjustment to backward pagination. If a count
is not provided, use 1 rather than 0, to advance to previous item.

Closes gh-929