Add empty "data:" field for SSE complete event
Prior to this commit, the "event:complete" event would have no "data:"
field when completing the stream of data. This can cause problems with
browsers, as the `EventSource` callback might not get called without it.

This commit ensures that an empty data field is present in all cases.

Fixes gh-940