Handle early GraphQLExceptions in ExecutionGraphQlService
Prior to this commit, the `ExecutionGraphQlService` would throw
`GrahphQLException` instances if errors happen early in the process, for
example in some cases of document parsing/validation.

Such exception would be thrown directly and not reflected in the GraphQL
response if those are `GraphQLError`.

This commit ensures that a fallback GraphQL response is created in this
case and that it contains the relevant `GraphQLError` information.

Fixes gh-1118